+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/activity/ApiMainActivity.kt
@@ -1,4 +1,4 @@
-package com.asiantech.intern20summer1.w10.activity
+package com.asiantech.intern20summer1.w11.activity
 
 import android.content.Context
 import android.os.Build
@@ -9,7 +9,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import com.asiantech.intern20summer1.R
-import com.asiantech.intern20summer1.w10.fragment.SplashFragment
+import com.asiantech.intern20summer1.w11.fragment.SplashFragment
 
 /**
  * Asian Tech Co., Ltd.
