+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/fragment/SignUpFragment.kt
@@ -1,4 +1,4 @@
-package com.asiantech.intern20summer1.w10.fragment
+package com.asiantech.intern20summer1.w11.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,12 +8,12 @@ import androidx.core.util.PatternsCompat
 import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.Fragment
 import com.asiantech.intern20summer1.R
-import com.asiantech.intern20summer1.w10.api.Api
-import com.asiantech.intern20summer1.w10.api.ApiAccountService
-import com.asiantech.intern20summer1.w10.api.ErrorUtils
-import com.asiantech.intern20summer1.w10.models.Account
-import com.asiantech.intern20summer1.w10.models.RequestAccount
-import com.asiantech.intern20summer1.w10.utils.AppUtils
+import com.asiantech.intern20summer1.w11.api.Api
+import com.asiantech.intern20summer1.w11.api.ApiAccountService
+import com.asiantech.intern20summer1.w11.api.ErrorUtils
+import com.asiantech.intern20summer1.w11.models.Account
+import com.asiantech.intern20summer1.w11.models.RequestAccount
+import com.asiantech.intern20summer1.w11.utils.AppUtils
 import kotlinx.android.synthetic.`at-huybui`.w10_fragment_sign_up.*
 import retrofit2.Call
 import retrofit2.Response
