+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/fragment/SplashFragment.kt
@@ -1,4 +1,4 @@
-package com.asiantech.intern20summer1.w10.fragment
+package com.asiantech.intern20summer1.w11.fragment
 
 import android.content.Context
 import android.content.Intent
@@ -14,12 +14,12 @@ import android.view.ViewGroup
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import com.asiantech.intern20summer1.R
-import com.asiantech.intern20summer1.w10.activity.ApiMainActivity
-import com.asiantech.intern20summer1.w10.api.Api
-import com.asiantech.intern20summer1.w10.api.ApiAccountService
-import com.asiantech.intern20summer1.w10.api.ErrorUtils
-import com.asiantech.intern20summer1.w10.models.Account
-import com.asiantech.intern20summer1.w10.utils.AppUtils
+import com.asiantech.intern20summer1.w11.activity.ApiMainActivity
+import com.asiantech.intern20summer1.w11.api.Api
+import com.asiantech.intern20summer1.w11.api.ApiAccountService
+import com.asiantech.intern20summer1.w11.api.ErrorUtils
+import com.asiantech.intern20summer1.w11.models.Account
+import com.asiantech.intern20summer1.w11.utils.AppUtils
 import retrofit2.Call
 import retrofit2.Response
