+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/utils/AppUtils.kt
@@ -1,4 +1,4 @@
-package com.asiantech.intern20summer1.w10.utils
+package com.asiantech.intern20summer1.w11.utils
 
 import android.content.Context
 import android.widget.Toast
